Update the static asset sync command
After every deployment we will sync (CSS, JS, images, etc) in the `wp-admin/` and `wp-includes/` directories as well as in the `themes/` and `plugins/` directories. We rewrite the URLs for these items so they are served from the CDN so we need to have a way to sync them to our Amazon S3 bucket.

`--no-mime-magic` and `--guess-mime-type` store the proper mime type for the item when it is synced

`--storage-class` tells Amazon S3 to use reduced redundancy which is slightly cheaper than standard storage. And if S3 should lose an item its no big deal since these are just copies from files in our repo.