Updated our uploaded media URLs
a.spirited.media points to BunnyCDN which is a far cheaper CDN for static assets. If an asset isn't found on BunyCDN it passes the request to CloudFront which is using a Lambda@Edge function to dynamically resize images. CloudFront pulls from our S3 bucket for the original uploaded media item.